requirements.txt
langchain
openai
pandas 
matplotlib
huggingface_hub
python-dotenv
streamlit